# server-mcp/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "私有化文档检索服务 - 为企业内网的 AI IDE 提供实时、准确的技术文档和代码示例",
        "title": "go-mcp-context API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "10.21.71.19:8090",
    "basePath": "/",
    "paths": {
...
        "/api/v1/documents/chunks/{mode}/{libid}": {
            "get": {
                "description": "获取指定库的文档块，支持向量搜索和全量列表。mode 为 code 或 info，version 可选（默认使用库的默认版本），topic 可选（传入则进行向量搜索）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "获取文档块",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模式：code(代码示例) 或 info(文档说明)",
                        "name": "mode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "库 ID",
                        "name": "libid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "版本号，不传则使用库的默认版本",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索主题，不传则返回全部文档块",
                        "name": "topic",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/detail/:id": {
            "get": {
                "description": "获取指定文档的完整信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "获取文档详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文档 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/list": {
            "get": {
                "description": "分页获取指定库的文档列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "获取文档列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库 ID",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "版本号（可选）",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量，默认 10，最大 50",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/upload": {
            "post": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "上传文档文件到指定库版本（支持 Markdown、PDF、DOCX、Swagger）（需要认证）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "上传文档",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库 ID",
                        "name": "library_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "版本号，默认为 latest",
                        "name": "version",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "文档文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/upload-sse": {
            "post": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "上传文档文件到指定库版本，通过 SSE 实时推送处理进度（需要认证）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "上传文档（SSE 实时推送）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库 ID",
                        "name": "library_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "版本号，默认为 latest",
                        "name": "version",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "文档文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "处理状态流",
                        "schema": {
                            "$ref": "#/definitions/response.ProcessStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ProcessStatus": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "progress": {
                    "description": "0-100",
                    "type": "integer"
                },
                "stage": {
                    "description": "uploaded, parsing, chunking, embedding, saving, completed, failed",
                    "type": "string"
                },
                "status": {
                    "description": "processing, active, failed",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTAuth": {
            "description": "JWT token in the format \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "MCP_API_KEY": {
            "description": "API Key for MCP protocol calls",
            "type": "apiKey",
            "name": "MCP_API_KEY",
            "in": "header"
        }
    }
}